Add new views for email preview and changes on utils and email task
